AI/app/requirements.txt
fastapi
uvicorn
pydantic
cohere
tiktoken
langchain
pinecone-client
langchain-google-genai
google-generativeai
langchain_community
langchain-experimental